+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Bug in json handling
+
+### Added
+
+- Health status endpoint `/info/health/status`.
+
 ## [0.3.2] - 2025-08-20
 
 ### Fixed